… src/docs/src/api/database/find.rst
… src/mango/src/mango_idx_text.erl
… src/mango/src/mango_idx_view.erl
… src/mango/src/mango_selector.erl
… src/mango/src/mango_selector_text.erl
… src/mango/test/22-covering-index-test.py