This pull request aims to implement the first phase of the covering indexes work (_"keys-only covering indexes"_) as it was described in the corresponding RFC document by @mikerhodes, see #4410 and #4413 for the details.

Ideally, end users should not notice any change in the behavior but speed-ups for certain kind of queries.  Performance results obtained by [k6](https://k6.io/) on my MacBook Pro 2021 M1 Max, with 10,000 documents, compared against `c5a7809f` with 3 nodes:

- 1-field index covers the query:
  - before: `min=0.59ms  max=498.29ms   p(95)=313.02ms  p(98)=357.44ms  p(99)=390.75ms`
  - after: `min=0.53ms max=127.26ms   p(95)=33.15ms   p(98)=37.53ms   p(99)=39.99ms`
- 2-field index covers the query:
  - before: `min=19.68ms max=893.79ms   p(95)=674.84ms  p(98)=739.12ms  p(99)=774.02ms`
  - after: `min=1.95ms max=163.22ms   p(95)=50.13ms   p(98)=57.11ms   p(99)=63.38ms`
- generic case:
  - before: `min=0.70ms  max=3786.70ms  p(95)=2733.86ms p(98)=3265.91ms p(99)=3452.13ms`
  - after: `min=0.72ms max=3590.26ms  p(95)=2631.55ms p(98)=2777.04ms p(99)=2879.50ms`

Please find the details of the performance evaluation in [`pgj/couchdb-k6/mango/couchdb_4482`](https://github.com/pgj/couchdb-k6/tree/main/mango/couchdb_4482).

Unit test coverage increased significantly, see before:

```console
$ make eunit apps=mango
==> mango (compile)
==> rel (compile)
==> couchdb (compile)
==> couchdb (setup_eunit)
==> mango (eunit)
======================== EUnit ========================
[..]
module 'mango_selector'
  mango_selector: is_constant_field_basic_test...ok
  mango_selector: is_constant_field_basic_two_test...ok
  mango_selector: is_constant_field_not_eq_test...ok
  mango_selector: is_constant_field_missing_field_test...ok
  mango_selector: is_constant_field_or_field_test...ok
  mango_selector: is_constant_field_empty_selector_test...ok
  mango_selector: is_constant_nested_and_test...ok
  mango_selector: is_constant_combined_or_and_equals_test...ok
  mango_selector: has_required_fields_basic_test...ok
  mango_selector: has_required_fields_basic_failure_test...ok
  mango_selector: has_required_fields_empty_selector_test...ok
  mango_selector: has_required_fields_exists_false_test...ok
  mango_selector: has_required_fields_and_true_test...ok
  mango_selector: has_required_fields_nested_and_true_test...ok
  mango_selector: has_required_fields_and_false_test...ok
  mango_selector: has_required_fields_or_false_test...ok
  mango_selector: has_required_fields_or_true_test...ok
  mango_selector: has_required_fields_and_nested_or_true_test...ok
  mango_selector: has_required_fields_and_nested_or_false_test...ok
  mango_selector: has_required_fields_or_nested_and_true_test...ok
  mango_selector: has_required_fields_or_nested_or_true_test...ok
  mango_selector: has_required_fields_or_nested_or_false_test...ok
  mango_selector:1013: match_demo_test_...[0.008 s] ok
  mango_selector:1014: match_demo_test_...ok
  mango_selector:1015: match_demo_test_...ok
  mango_selector:1017: match_demo_test_...ok
  mango_selector:1019: match_demo_test_...ok
  mango_selector:1020: match_demo_test_...ok
  [done in 0.066 s]
[..]
module 'mango_idx_view'
module 'mango_cursor_view'
  mango_cursor_view: does_not_refetch_doc_with_value_test...ok
  mango_cursor_view: doc_member_and_extract_fields_test...ok
  mango_cursor_view: match_and_extract_doc_match_test...ok
  mango_cursor_view: match_and_extract_doc_matchextract_test...ok
  mango_cursor_view: match_and_extract_doc_nomatch_test...ok
  mango_cursor_view: match_and_extract_doc_nomatch_fields_test...ok
  mango_cursor_view: choose_best_index_with_singleton_test...ok
  mango_cursor_view: choose_best_index_lowest_difference_test...ok
  mango_cursor_view: choose_best_index_least_number_of_fields_test...ok
  mango_cursor_view: choose_best_index_lowest_index_triple_test...ok
  [done in 0.030 s]
[..]
=======================================================
  All 81 tests passed.
[..]

Code Coverage:
mango_app             :   0%
mango_crud            :   0%
mango_cursor          :   0%
mango_cursor_special  :   0%
mango_cursor_text     :   0%
mango_cursor_view     :  33%
mango_doc             :   3%
mango_epi             : 100%
mango_error           :   0%
mango_execution_stats :   0%
mango_fields          :  50%
mango_httpd           :   0%
mango_httpd_handlers  :   0%
mango_idx             :  18%
mango_idx_special     :   0%
mango_idx_text        :  63%
mango_idx_view        :   1%
mango_json            :  10%
mango_json_bookmark   :   0%
mango_native_proc     :   5%
mango_opts            :  25%
mango_selector        :  62%
mango_selector_text   :  86%
mango_sort            :   0%
mango_sup             :   0%
mango_util            :  38%

Total                 : 31%
==> rel (eunit)
==> couchdb (eunit)
```

and then after:

```console
==> mango (compile)
==> rel (compile)
==> couchdb (compile)
==> couchdb (setup_eunit)
==> mango (eunit)
======================== EUnit ========================
[..]
module 'mango_selector'
  mango_selector: is_constant_field_basic_test...ok
  mango_selector: is_constant_field_basic_two_test...ok
  mango_selector: is_constant_field_not_eq_test...ok
  mango_selector: is_constant_field_missing_field_test...ok
  mango_selector: is_constant_field_or_field_test...ok
  mango_selector: is_constant_field_empty_selector_test...ok
  mango_selector: is_constant_nested_and_test...ok
  mango_selector: is_constant_combined_or_and_equals_test...ok
  mango_selector: has_required_fields_basic_test...ok
  mango_selector: has_required_fields_basic_failure_test...ok
  mango_selector: has_required_fields_empty_selector_test...ok
  mango_selector: has_required_fields_exists_false_test...ok
  mango_selector: has_required_fields_and_true_test...ok
  mango_selector: has_required_fields_nested_and_true_test...ok
  mango_selector: has_required_fields_and_false_test...ok
  mango_selector: has_required_fields_or_false_test...ok
  mango_selector: has_required_fields_or_true_test...ok
  mango_selector: has_required_fields_and_nested_or_true_test...ok
  mango_selector: has_required_fields_and_nested_or_false_test...ok
  mango_selector: has_required_fields_or_nested_and_true_test...ok
  mango_selector: has_required_fields_or_nested_or_true_test...ok
  mango_selector: has_required_fields_or_nested_or_false_test...ok
  mango_selector:1013: match_demo_test_...[0.008 s] ok
  mango_selector:1014: match_demo_test_...ok
  mango_selector:1015: match_demo_test_...ok
  mango_selector:1017: match_demo_test_...ok
  mango_selector:1019: match_demo_test_...ok
  mango_selector:1020: match_demo_test_...ok
  mango_selector: fields_empty_test...ok
  mango_selector: fields_primitive_test...ok
  mango_selector: fields_nested_test...ok
  mango_selector: fields_and_test...ok
  mango_selector: fields_or_test...ok
  mango_selector: fields_nor_test...ok
  [done in 0.110 s]
[..]
module 'mango_idx_view'
  mango_idx_view: indexable_fields_empty_test...ok
  mango_idx_view: indexable_fields_and_test...ok
  mango_idx_view: indexable_fields_or_test...ok
  mango_idx_view: indexable_fields_nor_test...ok
  mango_idx_view: indexable_fields_all_test...ok
  mango_idx_view: indexable_fields_elemMatch_test...ok
  mango_idx_view: indexable_fields_allMatch_test...ok
  mango_idx_view: indexable_fields_keyMapMatch_test...ok
  mango_idx_view: indexable_fields_in_test...ok
  mango_idx_view: indexable_fields_nin_test...ok
  mango_idx_view: indexable_fields_not_test...ok
  mango_idx_view: indexable_fields_lt_test...ok
  mango_idx_view: indexable_fields_lte_test...ok
  mango_idx_view: indexable_fields_eq_test...ok
  mango_idx_view: indexable_fields_ne_test...ok
  mango_idx_view: indexable_fields_gte_test...ok
  mango_idx_view: indexable_fields_gt_test...ok
  mango_idx_view: indexable_fields_mod_test...ok
  mango_idx_view: indexable_fields_regex_test...ok
  mango_idx_view: indexable_fields_exists_test...ok
  mango_idx_view: indexable_fields_type_test...ok
  mango_idx_view: indexable_fields_size_test...ok
  mango_idx_view: covers_all_fields_test...ok
  mango_idx_view: covers_all_docs_test...ok
  mango_idx_view: covers_empty_index_test...ok
  mango_idx_view: covers_regular_index_test...ok
  [done in 0.081 s]
module 'mango_cursor_view'
  mango_cursor_view: viewcbargs_test...ok
  mango_cursor_view: maybe_replace_max_json_test...ok
  mango_cursor_view: base_opts_test...ok
  mango_cursor_view: apply_opts_empty_test...ok
  mango_cursor_view: apply_opts_r_test...ok
  mango_cursor_view: apply_opts_conflicts_test...ok
  mango_cursor_view: apply_opts_sort_test...ok
  mango_cursor_view: apply_opts_stale_test...ok
  mango_cursor_view: apply_opts_stable_test...ok
  mango_cursor_view: apply_opts_update_test...ok
  mango_cursor_view: apply_opts_partition_test...[0.020 s] ok
  mango_cursor_view: consider_index_coverage_positive_test...ok
  mango_cursor_view: consider_index_coverage_negative_test...ok
  mango_cursor_view: derive_doc_from_index_test...ok
  mango_cursor_view: composite_indexes_test...ok
  mango_cursor_view: create_test...ok
  mango_cursor_view: explain_test...ok
  mango_cursor_view:944: execute_test_ (t_execute_empty)...[0.005 s] ok
  mango_cursor_view:945: execute_test_ (t_execute_ok_all_docs)...[0.003 s] ok
  mango_cursor_view:946: execute_test_ (t_execute_ok_query_view)...[0.003 s] ok
  mango_cursor_view:947: execute_test_ (t_execute_error)...[0.001 s] ok
  mango_cursor_view:1112: view_cb_test_ (t_view_cb_meta)...[0.002 s] ok
  mango_cursor_view:1113: view_cb_test_ (t_view_cb_row_matching_regular_doc)...[0.002 s] ok
  mango_cursor_view:1114: view_cb_test_ (t_view_cb_row_non_matching_regular_doc)...[0.002 s] ok
  mango_cursor_view:1115: view_cb_test_ (t_view_cb_row_null_doc)...[0.002 s] ok
  mango_cursor_view:1116: view_cb_test_ (t_view_cb_row_missing_doc_triggers_quorum_fetch)...[0.001 s] ok
  mango_cursor_view:1117: view_cb_test_ (t_view_cb_row_matching_covered_doc)...[0.002 s] ok
  mango_cursor_view:1118: view_cb_test_ (t_view_cb_row_non_matching_covered_doc)...[0.002 s] ok
  mango_cursor_view:1119: view_cb_test_ (t_view_cb_row_backwards_compatible)...[0.001 s] ok
  mango_cursor_view:1120: view_cb_test_ (t_view_cb_complete)...[0.003 s] ok
  mango_cursor_view:1121: view_cb_test_ (t_view_cb_ok)...[0.001 s] ok
  mango_cursor_view:1279: maybe_send_mango_ping_test_ (t_maybe_send_mango_ping_nop)...[0.001 s] ok
  mango_cursor_view:1280: maybe_send_mango_ping_test_ (t_maybe_send_mango_ping_happens)...[0.001 s] ok
  mango_cursor_view: ddocid_test...ok
  mango_cursor_view: is_design_doc_test...ok
  mango_cursor_view:1316: handle_message_test_ (t_handle_message_meta)...ok
  mango_cursor_view:1317: handle_message_test_ (t_handle_message_row_ok_above_limit)...[0.002 s] ok
  mango_cursor_view:1318: handle_message_test_ (t_handle_message_row_ok_at_limit)...ok
  mango_cursor_view:1319: handle_message_test_ (t_handle_message_row_ok_skip)...ok
  mango_cursor_view:1320: handle_message_test_ (t_handle_message_row_ok_triggers_quorum_fetch_match)...[0.648 s] ok
  mango_cursor_view:1321: handle_message_test_ (t_handle_message_row_ok_triggers_quorum_fetch_no_match)...[0.001 s] ok
  mango_cursor_view:1322: handle_message_test_ (t_handle_message_row_no_match)...ok
  mango_cursor_view:1323: handle_message_test_ (t_handle_message_row_error)...[0.020 s] ok
  mango_cursor_view:1324: handle_message_test_ (t_handle_message_execution_stats)...ok
  mango_cursor_view:1325: handle_message_test_ (t_handle_message_complete)...ok
  mango_cursor_view:1326: handle_message_test_ (t_handle_message_error)...ok
  mango_cursor_view: handle_all_docs_message_ddoc_test...ok
  mango_cursor_view: handle_all_docs_message_row_test...ok
  mango_cursor_view: handle_all_docs_message_regular_test...ok
  mango_cursor_view: does_not_refetch_doc_with_value_test...ok
  mango_cursor_view: doc_member_and_extract_fields_test...ok
  mango_cursor_view: match_and_extract_doc_match_test...ok
  mango_cursor_view: match_and_extract_doc_matchextract_test...ok
  mango_cursor_view: match_and_extract_doc_nomatch_test...ok
  mango_cursor_view: match_and_extract_doc_nomatch_fields_test...ok
  mango_cursor_view: choose_best_index_with_singleton_test...ok
  mango_cursor_view: choose_best_index_lowest_difference_test...ok
  mango_cursor_view: choose_best_index_least_number_of_fields_test...ok
  mango_cursor_view: choose_best_index_lowest_index_triple_test...ok
  mango_cursor_view: update_bookmark_keys_test...ok
  [done in 3.526 s]
[..]
=======================================================
  All 166 tests passed.
[..]

Code Coverage:
mango_app             :   0%
mango_crud            :   0%
mango_cursor          :  20%
mango_cursor_special  :   0%
mango_cursor_text     :   0%
mango_cursor_view     : 100%
mango_doc             :   4%
mango_epi             : 100%
mango_error           :   0%
mango_execution_stats :  63%
mango_fields          :  57%
mango_httpd           :   0%
mango_httpd_handlers  :   0%
mango_idx             :  21%
mango_idx_special     :   0%
mango_idx_text        :  63%
mango_idx_view        :  39%
mango_json            :  12%
mango_json_bookmark   :  25%
mango_native_proc     :   5%
mango_opts            :  27%
mango_selector        :  66%
mango_selector_text   :  86%
mango_sort            :   5%
mango_sup             :   0%

Total                 : 49%
==> rel (eunit)
==> couchdb (eunit)
```

Tasks:

- [x] Code is written and works correctly
- [x] Changes are covered by tests
- [x] Documentation changes were made in the `src/docs` folder