@pgj 2023-04-12
@willholley you are right, the statistics counters are not yet updated.  The reason for that is twofold: @mikerhodes has started to work on this earlier independently and I was not sure of his status until recently, then I wanted to have this piece out first and work on the counters in a separate pull request.  They have their own trouble with mixed-version clusters and I did not want to complicate this change further with those parts.  There are further pull requests planned along the lines of the corresponding RFC anyway.